requirements.txt
grpcio
grpcio-tools
selenium
requests
webdriver-manager
protobuf
pydantic
python-dotenv